+++ eggs/Naaya/eggs/Naaya/CHANGELOG.rst
@@ -1,5 +1,6 @@
 2.12.63 (unreleased)
 --------------------
+* using repoze.sendmail instead of zope.sendmail for queuing [simiamih]
 
 2.12.62 (2012-02-15)
 --------------------